add subtest helper functions
It's not the easiest thing in the world to generated subtest TAP.

Well, it kinda is, but generating subtest TAP that actually matches the
output of the *installed* copy of Test::More is...  exciting.

Here we add a couple helper functions to handle the not-so-fun parts of
it.  This is only being released at the moment in a trial release, so
this may change (e.g. these functions pushed into
TAP::SimpleOutput::Utils or the like).